trekker ut form-data-parsing til util, og fikser problem der isCombined har disabled columns og blir columns fra formdata undefined
